requirements.txt
# FinSightAI - Complete Requirements
# Production-ready dependencies with version constraints

# =============================================================================
# CORE API DEPENDENCIES
# =============================================================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# =============================================================================
# AI/ML DEPENDENCIES
# =============================================================================
# Vector embeddings and AI
langchain>=0.1.0
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.0

# Vector database
chromadb>=0.4.0
chroma-hnswlib>=0.7.0

# Data processing and ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# =============================================================================
# DATA INGESTION DEPENDENCIES
# =============================================================================
# Web scraping and RSS
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
lxml>=4.9.0

# Date/time handling
python-dateutil>=2.8.0

# =============================================================================
# UTILITIES
# =============================================================================
# Progress bars and utilities
tqdm>=4.66.0

# Environment variables
python-dotenv>=1.0.0

# Google Generative AI
google-genai

# Path handling (Python < 3.4 compatibility)
pathlib2>=2.3.0; python_version < "3.4"

# =============================================================================
# DEVELOPMENT DEPENDENCIES (Optional)
# =============================================================================
# Uncomment for development
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.7.0